refactor: break down gbq.py file to several smaller ones
