+++ pandas_gbq/exceptions.py
@@ -3,6 +3,70 @@
 # license that can be found in the LICENSE file.
 
 
+class DatasetCreationError(ValueError):
+    """
+    Raised when the create dataset method fails
+    """
+
+
+class InvalidColumnOrder(ValueError):
+    """
+    Raised when the provided column order for output
+    results DataFrame does not match the schema
+    returned by BigQuery.
+    """
+
+
+class InvalidIndexColumn(ValueError):
+    """
+    Raised when the provided index column for output
+    results DataFrame does not match the schema
+    returned by BigQuery.
+    """
+
+
+class InvalidPageToken(ValueError):
+    """
+    Raised when Google BigQuery fails to return,
+    or returns a duplicate page token.
+    """
+
+
+class InvalidSchema(ValueError):
+    """
+    Raised when the provided DataFrame does
+    not match the schema of the destination
+    table in BigQuery.
+    """
+
+    def __init__(self, message: str):
+        self._message = message
+
+    @property
+    def message(self) -> str:
+        return self._message
+
+
+class NotFoundException(ValueError):
+    """
+    Raised when the project_id, table or dataset provided in the query could
+    not be found.
+    """
+
+
+class TableCreationError(ValueError):
+    """
+    Raised when the create table method fails
+    """
+
+    def __init__(self, message: str):
+        self._message = message
+
+    @property
+    def message(self) -> str:
+        return self._message
+
+
 class GenericGBQException(ValueError):
     """
     Raised when an unrecognized Google API Error occurs.
